Completed take damage effect and determind is dead

+++ Assets/Script/Game Saving/CharacterSaveData.cs
@@ -23,7 +23,7 @@ namespace Horo
         public float zPosition;
 
         [Header("Resources")]
-        public float currentHealth;
+        public int currentHealth;
         public float currentStamina;
 
         [Header("Stats")]
